Prevent overflowing links in the story preview
Long links stretch the with of the preview when they overflow. The input
side of the edit screen shrinks to accommodate the overflow. This is not
ideal because it becomes harder and harder to see your input as it is
reduced in size.

By using minmax function we can be sure that the input size will be at
least 50% of the parent width.

Setting a min width on the input side is not all that is required
because it still allows the long link to overflow requiring that the
user scroll from left to right on the screen to view the entire preview.

So to improve on the minmax solution we also force long links to break
when ever they reach the container boundary.