Replace sleep with assertions that wait
Based on the PR feedback we don't want to have sleep statements in our
feature specs.

One solution is to implement a wait_until or similar alternatives
ourselves.

Howeve from this article
https://www.varvet.com/blog/why-wait_until-was-removed-from-capybara/ it
seems like an even better approach is to use assertion that cause
capybara to wait is the prefered approach.

This commit adds assertions that wait for items to appear or dissapear
and so the sleep behaviour is achieved without sleeping.